feat(language-server): allow to replace FQN on code assist
Use textEdit instead of insertText to specify the entire range of
characters that should be replaced when accepting the suggestion.
This allows to replace fully qualified names like "sap.m.Button" with
the correct tag name, to make for an easier search.

This also makes sure we replace the whole name even if the user requested
the code assist in the middle of it (instead of inserting it and keeping
the rest of the string): for example, selecting "MyClass2" when the
cursor is at "MyClass|1" will result in "MyClass2" instead of "MyClass21"
like before.

Bug fix: use prefix instead of the full attribute value in namespace
completion inside attribute value.