fix(language-server): detect the UI5 version from closest manifest.json (#472)
To support mono-repos having different UI5 projects and to ensure that
the correct minUI5Version of the UI5 application/library is being used
the UI5 version will now be derived from the closest manifest.json of
the XMLView to be edited.

`manifest.json`:

{
  [...]
  "sap.ui5": {
    "dependencies": {
      "minUI5Version": "${anySemVer}"
    }
  }
  [...]
}

The detection of the used framework has been removed so far and would
require an additional metadata in some UI5 application/library related
metadata. This can be re-added for future versions.