feat: add framework awareness by reading closest ui5.yaml
The ui5.yaml provides information about the UI5 framework to be
used for the application. There is no such information in the
manifest.json and therefore the ui5.yaml is the only indicator.
If no framework can be found in the ui5.yaml the default version
for now is SAPUI5.

Fixes #239